+++ src/core.cpp
@@ -1657,8 +1657,8 @@ void CUDT::CCUpdate()
    if (m_llMaxBW <= 0)
       return;
    const double minSP = 1000000.0 / (double(m_llMaxBW) / m_iMSS) * m_ullCPUFrequency;
-   if (m_ullInterval < minSP)
-       m_ullInterval = minSP;
+   if (m_ullInterval < (uint64_t)minSP)
+       m_ullInterval = (uint64_t)minSP;
 }
 
 void CUDT::initSynch()
